refactor(app): give the layout wiring's indicators and strip their own siblings
`app/layout_wiring.rs` carried 1,577 production lines across three
`impl QuantickApp` blocks. Two wiring groups move out unchanged:
`layout_wiring/indicators.rs` takes putting one indicator on one pane,
materialising a layout's saved set and mirroring an add, a removal, a
hide, an input edit or a style edit onto every pane showing the same
layout; `layout_wiring/strip.rs` takes the layout strip and the delete
confirmation it opens.

Pure moves: every body is unchanged. A method that was `pub(super)` from
this file keeps exactly the visibility it had, spelled `pub(in
crate::app)` from one level deeper -- the same visibility, not a
widening -- and one private helper the parent still calls becomes
`pub(super)`. The strip's two adjacent `impl QuantickApp` blocks become
one in the new module, which pays back the `impl`/`}` pair the
indicators module adds, so the extension-boundary root count falls
rather than rises.
