+++ crates/app/src/app/layout_wiring.rs
@@ -54,6 +54,9 @@ use super::chrome::ChromeState;
 use super::{QuantickApp, TabSlot};
 use crate::workspace_store::LayoutSave;
 
+mod indicators;
+mod strip;
+
 /// The feed half of a drawing key while a tab plays a recording.
 ///
 /// A recording is its own market: its prices are the recorded day's, not
@@ -619,463 +622,6 @@ impl QuantickApp {
             .collect()
     }
 
-    /// Where a slot sits in its pane's layout, or `None` for a slot the
-    /// layout does not carry — an operator's, or one a validation hook added.
-    fn layout_index_of(&self, target: TabSlot) -> Option<usize> {
-        self.layout_slots_at(target.tab, target.side)
-            .iter()
-            .position(|slot| *slot == target.slot)
-    }
-
-    /// Add one indicator to one pane, with no fan-out and no dirty mark —
-    /// the primitive both the mirror and the materialisation are built on.
-    pub(super) fn add_indicator_at(
-        &mut self,
-        tab: u64,
-        side: PaneSide,
-        kind: &SavedKind,
-    ) -> Option<SlotId> {
-        let source = match kind {
-            // Unresolved on purpose: the worker owns the catalog, and an id
-            // it does not know becomes an error slot naming it.
-            SavedKind::Native { id } => IndicatorSource::Native {
-                id: id.clone(),
-                values: Vec::new(),
-            },
-            SavedKind::Script { name } => {
-                let index = self
-                    .indicators
-                    .script_library
-                    .entries()
-                    .iter()
-                    .position(|candidate| candidate.name == *name);
-                // A script the library no longer has still takes a slot — an
-                // error slot saying so. The layout addresses its panes by
-                // entry index, and a pane with one slot fewer than its layout
-                // has entries would have every edit after the gap land one
-                // entry off; and a row that says "not in the library" is the
-                // honest picture of a set the trader cannot see whole.
-                let read = match index {
-                    Some(index) => self.indicators.script_library.read(index),
-                    None => {
-                        tracing::warn!(
-                            target: "quantick::app",
-                            schema_version = 1_u8,
-                            event_code = "INDICATOR_STATE_SCRIPT_MISSING",
-                            script = %name,
-                            action = "error_slot_shown",
-                            "the layout references a script the library no longer has"
-                        );
-                        Some(Err(format!("{name} is not in the script library")))
-                    }
-                };
-                let file_info =
-                    index.and_then(|index| self.indicators.script_library.file_info(index));
-                let slot = match read {
-                    Some(Ok(text)) => {
-                        let pane = self.pane_mut_at(tab, side)?;
-                        pane.add_indicator(IndicatorSource::Script {
-                            name: name.clone(),
-                            text,
-                        })
-                    }
-                    Some(Err(message)) => {
-                        if index.is_some() {
-                            tracing::warn!(
-                                target: "quantick::app",
-                                schema_version = 1_u8,
-                                event_code = "INDICATOR_SCRIPT_UNREADABLE",
-                                script = %name,
-                                error = %message,
-                                action = "error_slot_shown",
-                                "cannot read an indicator script"
-                            );
-                        }
-                        // The same error slot the menu's own add builds, so a
-                        // script the trader fixes and reloads keeps its place.
-                        let pane = self.pane_mut_at(tab, side)?;
-                        let slot = pane.indicators.allocate_slot(format!("script.{name}"));
-                        pane.indicators.apply(IndicatorEvent::Rebuilt {
-                            slot,
-                            descriptor: quantick_indicators::IndicatorDescriptor {
-                                title: name.clone(),
-                                short_title: None,
-                                overlay: false,
-                                plots: Vec::new(),
-                                fills: Vec::new(),
-                                inputs: Vec::new(),
-                            },
-                            columns: Vec::new(),
-                            bar_paint: Vec::new(),
-                            rows: 0,
-                            inputs: Vec::new(),
-                            stale: None,
-                        });
-                        pane.indicators.apply(IndicatorEvent::Error {
-                            slot,
-                            error: quantick_indicators::EvalError {
-                                bar_index: 0,
-                                message,
-                            },
-                        });
-                        slot
-                    }
-                    None => return None,
-                };
-                let owner = TabSlot { tab, side, slot };
-                self.indicators.slot_kinds.push((owner, kind.clone()));
-                if let (Some(index), Some((_, mtime))) = (index, file_info) {
-                    self.indicators.script_files.push((owner, index, mtime));
-                }
-                return Some(slot);
-            }
-        };
-        let slot = self.pane_mut_at(tab, side)?.add_indicator(source);
-        self.indicators
-            .slot_kinds
-            .push((TabSlot { tab, side, slot }, kind.clone()));
-        Some(slot)
-    }
-
-    /// Take every layout slot off one pane, view and worker alike.
-    fn remove_layout_indicators_at(&mut self, tab: u64, side: PaneSide) {
-        for slot in self.layout_slots_at(tab, side) {
-            self.remove_indicator_silently(TabSlot { tab, side, slot });
-        }
-    }
-
-    /// Remove one slot with no fan-out and no dirty mark.
-    fn remove_indicator_silently(&mut self, target: TabSlot) {
-        let pane = self
-            .tabs
-            .iter_mut()
-            .find(|tab| tab.id == target.tab)
-            .and_then(|tab| tab.pane_at_mut(target.side.index()));
-        self.indicators.slots_mut().remove(pane, target);
-    }
-
-    /// Put a whole saved set on one pane: add, bind inputs, queue the hide
-    /// and the style until the worker's first `Rebuilt` gives them a view.
-    pub(super) fn materialize_indicators_at(
-        &mut self,
-        tab: u64,
-        side: PaneSide,
-        set: &[SavedIndicator],
-    ) {
-        for entry in set {
-            let Some(slot) = self.add_indicator_at(tab, side, &entry.kind) else {
-                continue;
-            };
-            let owner = TabSlot { tab, side, slot };
-            self.apply_saved_entry(owner, entry);
-        }
-    }
-
-    /// Bind one saved entry's inputs, hide flag and style onto a slot.
-    fn apply_saved_entry(&mut self, owner: TabSlot, entry: &SavedIndicator) {
-        let values: Vec<_> = entry
-            .inputs
-            .iter()
-            .filter_map(SavedInput::to_value)
-            .collect();
-        if !values.is_empty() && values.len() == entry.inputs.len() {
-            if let Some(pane) = self.pane_mut_at(owner.tab, owner.side) {
-                pane.indicator_worker.send(IndicatorCommand::SetInputs {
-                    slot: owner.slot,
-                    values,
-                });
-            }
-        } else if !entry.inputs.is_empty() {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "INDICATOR_STATE_INPUTS_DROPPED",
-                kind = ?entry.kind,
-                saved = entry.inputs.len(),
-                readable = values.len(),
-                action = "declared_defaults_used",
-                "saved indicator inputs could not be read; using the declared defaults"
-            );
-        }
-        if entry.hidden {
-            self.indicators.pending_hidden.push(owner);
-        }
-        if !entry.plot_styles.is_empty() {
-            self.indicators.pending_styles.push((
-                owner,
-                StyleOverride::from_plots(
-                    entry
-                        .plot_styles
-                        .iter()
-                        .copied()
-                        .map(SavedPlotStyle::to_override)
-                        .collect(),
-                ),
-            ));
-        }
-    }
-
-    /// Apply queued hide flags and styles to the views that now exist.
-    ///
-    /// A queued hide means "hide this once it is born": views are born
-    /// visible, so applying it is a toggle. Whoever queues one dequeues it
-    /// again on the matching unhide ([`Self::mirror_hidden`]).
-    pub(super) fn apply_pending_indicator_state(&mut self) {
-        if !self.indicators.pending_hidden.is_empty() {
-            let pending = std::mem::take(&mut self.indicators.pending_hidden);
-            for owner in pending {
-                let applied = self
-                    .pane_mut_at(owner.tab, owner.side)
-                    .filter(|pane| {
-                        pane.indicators
-                            .all()
-                            .iter()
-                            .any(|view| view.slot == owner.slot)
-                    })
-                    .map(|pane| pane.indicators.toggle_hidden(owner.slot))
-                    .is_some();
-                if !applied
-                    && self
-                        .indicators
-                        .slot_kinds
-                        .iter()
-                        .any(|(candidate, _)| *candidate == owner)
-                {
-                    self.indicators.pending_hidden.push(owner);
-                }
-            }
-        }
-        if !self.indicators.pending_styles.is_empty() {
-            let pending = std::mem::take(&mut self.indicators.pending_styles);
-            for (owner, style) in pending {
-                let known = self
-                    .indicators
-                    .slot_kinds
-                    .iter()
-                    .any(|(candidate, _)| *candidate == owner);
-                match self
-                    .pane_mut_at(owner.tab, owner.side)
-                    .and_then(|pane| pane.indicators.view_mut(owner.slot))
-                {
-                    Some(view) => view.style = style,
-                    None if known => self.indicators.pending_styles.push((owner, style)),
-                    None => {}
-                }
-            }
-        }
-    }
-
-    /// The origin pane's layout and the index of the slot in it — the two
-    /// coordinates every mirror writes to.
-    fn edit_coordinates(&self, origin: TabSlot) -> Option<(LayoutId, usize)> {
-        let index = self.layout_index_of(origin)?;
-        Some((self.pane_layout(origin.tab, origin.side), index))
-    }
-
-    /// A layout's entry at a layout index, for an edit to write.
-    fn layout_entry_mut(&mut self, layout: LayoutId, index: usize) -> Option<&mut SavedIndicator> {
-        self.workspace
-            .layouts_mut()
-            .book_mut()
-            .get_mut(layout)?
-            .indicators
-            .get_mut(index)
-    }
-
-    /// The other panes an edit on `origin` reaches: every pane on the same
-    /// layout, minus the origin.
-    fn mirror_targets(&self, origin: TabSlot, layout: LayoutId) -> Vec<(u64, PaneSide)> {
-        self.panes_on(layout)
-            .into_iter()
-            .filter(|target| *target != (origin.tab, origin.side))
-            .collect()
-    }
-
-    /// Mirror an add: the layout gains the entry, and the same kind goes on
-    /// every other pane showing it, so the new indicator is on every such
-    /// chart the frame it was asked for. Inputs start empty — "the declared
-    /// defaults" — until the trader commits some.
-    /// The caller marks save intent once after completing the origin edit.
-    pub(super) fn mirror_add(&mut self, origin: TabSlot, kind: &SavedKind) {
-        let Some((layout, index)) = self.edit_coordinates(origin) else {
-            return;
-        };
-        let entry = SavedIndicator {
-            kind: kind.clone(),
-            hidden: false,
-            inputs: Vec::new(),
-            plot_styles: Vec::new(),
-        };
-        if let Some(target) = self.workspace.layouts_mut().book_mut().get_mut(layout) {
-            if index <= target.indicators.len() {
-                target.indicators.insert(index, entry);
-            } else {
-                target.indicators.push(entry);
-            }
-        }
-        for (tab, side) in self.mirror_targets(origin, layout) {
-            self.add_indicator_at(tab, side, kind);
-        }
-    }
-
-    /// Mirror a removal by layout index: the entry goes, and the slot at that
-    /// position on every other pane of the layout with it.
-    /// The caller marks save intent once after removing the origin.
-    pub(super) fn mirror_remove(&mut self, origin: TabSlot) {
-        let Some((layout, index)) = self.edit_coordinates(origin) else {
-            return;
-        };
-        if let Some(target) = self.workspace.layouts_mut().book_mut().get_mut(layout)
-            && index < target.indicators.len()
-        {
-            target.indicators.remove(index);
-        }
-        for (tab, side) in self.mirror_targets(origin, layout) {
-            if let Some(slot) = self.layout_slots_at(tab, side).get(index).copied() {
-                self.remove_indicator_silently(TabSlot { tab, side, slot });
-            }
-        }
-    }
-
-    /// Mirror an eye toggle: the entry records it, and the same position on
-    /// every other pane of the layout follows — now, or once its view is born.
-    pub(super) fn mirror_hidden(&mut self, origin: TabSlot) {
-        let Some((layout, index)) = self.edit_coordinates(origin) else {
-            return;
-        };
-        let Some(hidden) = self
-            .pane_at(origin.tab, origin.side)
-            .and_then(|pane| {
-                pane.indicators
-                    .all()
-                    .iter()
-                    .find(|view| view.slot == origin.slot)
-            })
-            .map(|view| view.hidden)
-        else {
-            return;
-        };
-        if let Some(entry) = self.layout_entry_mut(layout, index) {
-            entry.hidden = hidden;
-        }
-        for (tab, side) in self.mirror_targets(origin, layout) {
-            let Some(slot) = self.layout_slots_at(tab, side).get(index).copied() else {
-                continue;
-            };
-            let owner = TabSlot { tab, side, slot };
-            let Some(pane) = self.pane_mut_at(tab, side) else {
-                continue;
-            };
-            let live = pane
-                .indicators
-                .all()
-                .iter()
-                .find(|view| view.slot == slot)
-                .map(|view| view.hidden);
-            match live {
-                Some(live) if live != hidden => pane.indicators.toggle_hidden(slot),
-                Some(_) => {}
-                None => {
-                    // Unborn: the queue says what its first frame should
-                    // be. An unhide takes a queued hide back out, or the
-                    // view would be born and hidden a moment later.
-                    self.indicators
-                        .pending_hidden
-                        .retain(|candidate| *candidate != owner);
-                    if hidden {
-                        self.indicators.pending_hidden.push(owner);
-                    }
-                }
-            }
-        }
-        self.mark_layouts_dirty();
-    }
-
-    /// Mirror committed inputs: the entry records the values the trader
-    /// applied, and the same position on every other pane of the layout is
-    /// sent them.
-    ///
-    /// Committed, never previewed: the settings dialog's live preview goes
-    /// to the origin's worker alone and never reaches here, so a slider
-    /// mid-drag can never land in the file or on another chart.
-    pub(super) fn mirror_inputs(
-        &mut self,
-        origin: TabSlot,
-        values: &[quantick_indicators::InputValue],
-    ) {
-        let Some((layout, index)) = self.edit_coordinates(origin) else {
-            return;
-        };
-        if let Some(entry) = self.layout_entry_mut(layout, index) {
-            entry.inputs = values.iter().map(SavedInput::from_value).collect();
-        }
-        for (tab, side) in self.mirror_targets(origin, layout) {
-            if let Some(slot) = self.layout_slots_at(tab, side).get(index).copied()
-                && let Some(pane) = self.pane_mut_at(tab, side)
-            {
-                pane.indicator_worker.send(IndicatorCommand::SetInputs {
-                    slot,
-                    values: values.to_vec(),
-                });
-            }
-        }
-        self.mark_layouts_dirty();
-    }
-
-    /// Mirror the origin's style layer: the entry records it, and the same
-    /// position on every other pane of the layout wears it.
-    pub(super) fn mirror_style(&mut self, origin: TabSlot) {
-        let Some((layout, index)) = self.edit_coordinates(origin) else {
-            return;
-        };
-        let Some(style) = self
-            .pane_at(origin.tab, origin.side)
-            .and_then(|pane| {
-                pane.indicators
-                    .all()
-                    .iter()
-                    .find(|view| view.slot == origin.slot)
-            })
-            .map(|view| view.style.clone())
-        else {
-            return;
-        };
-        if let Some(entry) = self.layout_entry_mut(layout, index) {
-            entry.plot_styles = style
-                .plots()
-                .iter()
-                .copied()
-                .map(SavedPlotStyle::from_override)
-                .collect();
-        }
-        for (tab, side) in self.mirror_targets(origin, layout) {
-            let Some(slot) = self.layout_slots_at(tab, side).get(index).copied() else {
-                continue;
-            };
-            let owner = TabSlot { tab, side, slot };
-            match self
-                .pane_mut_at(tab, side)
-                .and_then(|pane| pane.indicators.view_mut(slot))
-            {
-                Some(view) => view.style = style.clone(),
-                None => {
-                    self.indicators
-                        .pending_styles
-                        .retain(|(candidate, _)| *candidate != owner);
-                    self.indicators.pending_styles.push((owner, style.clone()));
-                }
-            }
-        }
-        self.mark_layouts_dirty();
-    }
-
-    /// An indicator edit happened on a pane. The mirror that made it has
-    /// already written the layout; this starts the save clock for edits
-    /// that reach the layout by no other path.
-    pub(super) fn note_indicator_edit_at(&mut self, _tab: u64, _side: PaneSide) {
-        self.mark_layouts_dirty();
-    }
-
     /// Panes that appeared since last frame — a tab opened, a split built —
     /// take a layout: the one a restored workspace named for them, else the
     /// focused pane's, else the book's default. Its indicators go on, and
@@ -1347,219 +893,6 @@ impl QuantickApp {
                     twin.drawings.select(index);
                 }
             }
-        }
-    }
-}
-
-impl QuantickApp {
-    // ------------------------------------------------------------------
-    // Reload, rename, strip
-    // ------------------------------------------------------------------
-
-    /// Read the layouts file again and put it on every pane — after a
-    /// workspace import replaced the file under the running app. Every pane
-    /// reopens on the layout the imported workspace named for it, else the
-    /// book's default.
-    ///
-    /// `imported` names the stores the import wrote. A bundle from before
-    /// layouts existed carries an `indicators` section and no `layouts`
-    /// section; reading the cockpit's own layouts file back would show the
-    /// cockpit's indicators under a toast saying the bundle's were restored,
-    /// so that case migrates the imported set into "Layout 1" instead —
-    /// the same migration a launch performs.
-    pub(super) fn reload_layouts(&mut self, imported: &[&str]) {
-        self.clear_indicators();
-        for tab in &mut self.tabs {
-            for pane in tab.panes_mut() {
-                pane.drawings.take_all();
-                pane.drawings_key = None;
-                pane.layout_seeded = false;
-            }
-        }
-        let legacy = crate::indicators::state_file::default_path();
-        let migrate_imported_indicators =
-            imported.contains(&"indicators") && !imported.contains(&"layouts");
-        let (book, blocked) = if migrate_imported_indicators {
-            let set = crate::indicators::state_file::load(&legacy);
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "LAYOUTS_MIGRATED",
-                indicators = set.len(),
-                action = "imported_indicator_state_moved_into_layout_1",
-                "the imported bundle predates layouts; its indicator set became Layout 1"
-            );
-            (LayoutBook::starter(set), false)
-        } else {
-            Self::load_layouts(self.workspace.layouts_path(), &legacy)
-        };
-        self.workspace.layouts_mut().set_book(book);
-        self.workspace.layouts_mut().set_blocked(blocked);
-        self.chrome.layout_rename = None;
-        self.chrome.layout_delete_confirm = None;
-        // A pane whose named layout is not in the new book opens on the
-        // default, exactly as a pane with no name would.
-        for tab in &mut self.tabs {
-            for pane in tab.panes_mut() {
-                let book = self.workspace.layouts().book();
-                if pane.layout.is_some_and(|id| book.get(id).is_none()) {
-                    pane.layout = None;
-                }
-            }
-        }
-        self.seed_new_panes();
-        // Last, after the seeding, and `LayoutStore::settle` says why.
-        self.workspace
-            .layouts_mut()
-            .settle(migrate_imported_indicators, Instant::now());
-    }
-
-    /// Open the rename box on `id`, seeded with its current name.
-    pub(crate) fn begin_layout_rename(&mut self, id: LayoutId) {
-        if let Some(layout) = self.layouts().get(id) {
-            self.chrome.layout_rename = Some((id, layout.name.clone()));
-        }
-    }
-
-    /// Draw the strip and apply what it asked for.
-    pub(super) fn draw_layout_strip(&mut self, ctx: &eframe::egui::Context) {
-        let actions = {
-            let can_add = self.layouts().layouts().len() < layouts::MAX_LAYOUTS;
-            let can_delete = self.layouts().layouts().len() > 1;
-            let active = self.focused_pane_layout();
-            let owner = self.active_tab().focused_side().title();
-            let Self {
-                chrome: ChromeState { layout_rename, .. },
-                workspace,
-                ..
-            } = self;
-            crate::layout_strip::draw(
-                ctx,
-                crate::layout_strip::StripModel {
-                    layouts: workspace.layouts().book().layouts(),
-                    active,
-                    owner: &owner,
-                    rename: layout_rename,
-                    can_add,
-                    can_delete,
-                },
-            )
-        };
-        for action in actions {
-            self.apply_strip_action(action);
-        }
-    }
-
-    /// One door for the strip, the menu and the keyboard's rename box.
-    pub(crate) fn apply_strip_action(&mut self, action: crate::layout_strip::StripAction) {
-        use crate::layout_strip::StripAction;
-        let outcome: Result<(), LayoutError> = match action {
-            StripAction::Switch(id) => self.switch_layout(id).map(|_| ()),
-            StripAction::Create => self.create_layout(None).map(|_| ()),
-            StripAction::BeginRename(id) => {
-                self.begin_layout_rename(id);
-                Ok(())
-            }
-            StripAction::CommitRename(id, name) => {
-                self.chrome.layout_rename = None;
-                // An empty box is a cancelled rename, not an error to show.
-                match layouts::clean_name(&name) {
-                    Some(_) => self.rename_layout(id, &name).map(|_| ()),
-                    None => Ok(()),
-                }
-            }
-            StripAction::CancelRename => {
-                self.chrome.layout_rename = None;
-                Ok(())
-            }
-            // Deleting takes the layout's drawings with it, on disk as well
-            // as on screen: the one strip action that asks first.
-            StripAction::Delete(id) => {
-                if self.layouts().get(id).is_some() {
-                    self.chrome.layout_delete_confirm = Some(id);
-                }
-                Ok(())
-            }
-        };
-        if let Err(error) = outcome {
-            self.note_workspace(error.to_string());
-        }
-    }
-}
-
-impl QuantickApp {
-    /// The confirmation a delete waits on: the layout's name, what goes with
-    /// it, and the two buttons. Enter deletes and Escape cancels — both
-    /// consumed here, so neither reaches the chart's own key handling —
-    /// and nothing else on the window is touched while it is up.
-    pub(super) fn draw_layout_delete_confirm(&mut self, ctx: &eframe::egui::Context) {
-        use eframe::egui;
-        let Some(id) = self.chrome.layout_delete_confirm else {
-            return;
-        };
-        let Some(layout) = self.layouts().get(id) else {
-            self.chrome.layout_delete_confirm = None;
-            return;
-        };
-        let name = layout.name.clone();
-        let drawings: usize = layout.drawing_count();
-        // Counted, not collected: the window repaints continuously and only
-        // the number is read.
-        let showing = self.panes_on_count(id);
-        let mut decision: Option<bool> = None;
-        egui::Window::new("Delete layout")
-            .id(egui::Id::new("layout_delete_confirm"))
-            .collapsible(false)
-            .resizable(false)
-            .anchor(egui::Align2::CENTER_CENTER, egui::vec2(0.0, 0.0))
-            .show(ctx, |ui| {
-                ui.label(format!("Delete \"{name}\"?"));
-                let mut what = if drawings == 0 {
-                    "Its indicator set goes with it. Nothing is drawn under it.".to_owned()
-                } else {
-                    format!(
-                        "Its indicator set and the {drawings} drawing(s) kept under it go with it. This cannot be undone."
-                    )
-                };
-                if showing > 0 {
-                    what.push_str(&format!(
-                        " {showing} chart(s) showing it move to the layout beside it."
-                    ));
-                }
-                ui.label(
-                    egui::RichText::new(what)
-                        .small()
-                        .color(crate::theme::TEXT_SUPPORT),
-                );
-                ui.horizontal(|ui| {
-                    if ui.button("Delete").clicked() {
-                        decision = Some(true);
-                    }
-                    if ui.button("Cancel").clicked() {
-                        decision = Some(false);
-                    }
-                });
-            });
-        if ctx.input_mut(|input| input.consume_key(egui::Modifiers::NONE, egui::Key::Escape)) {
-            decision = Some(false);
-        } else if ctx.input_mut(|input| input.consume_key(egui::Modifiers::NONE, egui::Key::Enter))
-        {
-            decision = Some(true);
-        }
-        match decision {
-            Some(true) => self.confirm_layout_delete(),
-            Some(false) => self.chrome.layout_delete_confirm = None,
-            None => {}
-        }
-    }
-
-    /// The confirmed half of a delete: what the dialog's Delete button does.
-    pub(crate) fn confirm_layout_delete(&mut self) {
-        let Some(id) = self.chrome.layout_delete_confirm.take() else {
-            return;
-        };
-        if let Err(error) = self.delete_layout(id) {
-            self.note_workspace(error.to_string());
         }
     }
 }
